requirements.txt
python-nmap
get-mac
